Fixed discount on edit form

+++ db/migrate/20160718165206_add_discount_to_products.rb
@@ -1,5 +1,5 @@
 class AddDiscountToProducts < ActiveRecord::Migration
   def change
-    add_column :products, :discount, :float, default: 0
+    add_column :products, :discount, :decimal, default: 0
   end
 end
